+++ index.html
@@ -195,8 +195,6 @@
       </ul>
     </div>
   </footer>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 
   <!-- Шаблон изображения случайного пользователя -->
   <template id="picture">
@@ -235,6 +233,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-
+    <script src="js/functions.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
